+++ app/src/main/java/org/linphone/ui/main/chat/fragment/ConversationsListFragment.kt
@@ -33,7 +33,9 @@ import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
+import org.linphone.LinphoneApplication.Companion.coreContext
 import org.linphone.R
+import org.linphone.contacts.getListOfSipAddressesAndPhoneNumbers
 import org.linphone.core.tools.Log
 import org.linphone.databinding.ChatListFragmentBinding
 import org.linphone.ui.fileviewer.FileViewerActivity
@@ -41,10 +43,15 @@ import org.linphone.ui.fileviewer.MediaViewerActivity
 import org.linphone.ui.main.MainActivity.Companion.ARGUMENTS_CONVERSATION_ID
 import org.linphone.ui.main.chat.adapter.ConversationsListAdapter
 import org.linphone.ui.main.chat.viewmodel.ConversationsListViewModel
+import org.linphone.ui.main.contacts.model.ContactNumberOrAddressClickListener
+import org.linphone.ui.main.contacts.model.ContactNumberOrAddressModel
+import org.linphone.ui.main.contacts.model.NumberOrAddressPickerDialogModel
 import org.linphone.ui.main.fragment.AbstractMainFragment
 import org.linphone.ui.main.history.fragment.HistoryMenuDialogFragment
+import org.linphone.utils.DialogUtils
 import org.linphone.utils.Event
 import org.linphone.utils.LinphoneUtils
+import org.linphone.utils.RecyclerViewHeaderDecoration
 
 @UiThread
 class ConversationsListFragment : AbstractMainFragment() {
@@ -59,6 +66,21 @@ class ConversationsListFragment : AbstractMainFragment() {
     private lateinit var adapter: ConversationsListAdapter
 
     private var bottomSheetDialog: BottomSheetDialogFragment? = null
+
+    private val numberOrAddressClickListener = object : ContactNumberOrAddressClickListener {
+        @UiThread
+        override fun onClicked(model: ContactNumberOrAddressModel) {
+            coreContext.postOnCoreThread {
+                val address = model.address
+                if (address != null) {
+                    Log.i("$TAG Creating 1-1 conversation with to [${address.asStringUriOnly()}]")
+                    listViewModel.createOneToOneChatRoomWith(address)
+                }
+            }
+        }
+
+        override fun onLongPress(model: ContactNumberOrAddressModel) { }
+    }
 
     private val dataObserver = object : RecyclerView.AdapterDataObserver() {
         override fun onItemRangeInserted(positionStart: Int, itemCount: Int) {
@@ -121,6 +143,9 @@ class ConversationsListFragment : AbstractMainFragment() {
         binding.conversationsList.outlineProvider = outlineProvider
         binding.conversationsList.clipToOutline = true
 
+        val headerItemDecoration = RecyclerViewHeaderDecoration(requireContext(), adapter)
+        binding.conversationsList.addItemDecoration(headerItemDecoration)
+
         adapter.conversationLongClickedEvent.observe(viewLifecycleOwner) {
             it.consume { model ->
                 val modalBottomSheet = ConversationDialogFragment(
@@ -165,6 +190,35 @@ class ConversationsListFragment : AbstractMainFragment() {
             }
         }
 
+        adapter.createConversationWithFriendClickedEvent.observe(viewLifecycleOwner) {
+            it.consume { friend ->
+                coreContext.postOnCoreThread {
+                    val singleAvailableAddress = LinphoneUtils.getSingleAvailableAddressForFriend(friend)
+                    if (singleAvailableAddress != null) {
+                        Log.i(
+                            "$TAG Only 1 SIP address or phone number found for contact [${friend.name}], using it"
+                        )
+                        listViewModel.createOneToOneChatRoomWith(singleAvailableAddress)
+                    } else {
+                        val list = friend.getListOfSipAddressesAndPhoneNumbers(numberOrAddressClickListener)
+                        Log.i(
+                            "$TAG [${list.size}] numbers or addresses found for contact [${friend.name}], showing selection dialog"
+                        )
+                        coreContext.postOnMainThread {
+                            showNumbersOrAddressesDialog(list)
+                        }
+                    }
+                }
+            }
+        }
+
+        adapter.createConversationWithAddressClickedEvent.observe(viewLifecycleOwner) {
+            it.consume { address ->
+                Log.i("$TAG Creating 1-1 conversation with to [${address.asStringUriOnly()}]")
+                listViewModel.createOneToOneChatRoomWith(address)
+            }
+        }
+
         binding.setOnNewConversationClicked {
             if (findNavController().currentDestination?.id == R.id.conversationsListFragment) {
                 Log.i("$TAG Navigating to start conversation fragment")
@@ -185,6 +239,14 @@ class ConversationsListFragment : AbstractMainFragment() {
 
             Log.i("$TAG Conversations list ready with [${it.size}] items")
             listViewModel.fetchInProgress.value = false
+        }
+
+        listViewModel.chatRoomCreatedEvent.observe(viewLifecycleOwner) {
+            it.consume { conversationId ->
+                Log.i("$TAG Conversation [$conversationId] has been created, navigating to it")
+                val action = ConversationFragmentDirections.actionGlobalConversationFragment(conversationId)
+                binding.chatNavContainer.findNavController().navigate(action)
+            }
         }
 
         sharedViewModel.showConversationEvent.observe(viewLifecycleOwner) {
@@ -251,10 +313,10 @@ class ConversationsListFragment : AbstractMainFragment() {
 
         sharedViewModel.updateConversationLastMessageEvent.observe(viewLifecycleOwner) {
             it.consume { conversationId ->
-                val model = listViewModel.conversations.value.orEmpty().find { conversationModel ->
-                    conversationModel.id == conversationId
+                val model = listViewModel.conversations.value.orEmpty().find { wrapperModel ->
+                    wrapperModel.conversationModel?.id == conversationId
                 }
-                model?.updateLastMessageInfo()
+                model?.conversationModel?.updateLastMessageInfo()
             }
         }
 
@@ -262,10 +324,10 @@ class ConversationsListFragment : AbstractMainFragment() {
             it.consume {
                 val displayChatRoom = sharedViewModel.displayedChatRoom
                 if (displayChatRoom != null) {
-                    val found = listViewModel.conversations.value.orEmpty().find { model ->
-                        model.chatRoom == displayChatRoom
+                    val found = listViewModel.conversations.value.orEmpty().find { wrapperModel ->
+                        wrapperModel.conversationModel?.chatRoom == displayChatRoom
                     }
-                    found?.updateMuteState()
+                    found?.conversationModel?.updateMuteState()
                 }
             }
         }
@@ -277,9 +339,9 @@ class ConversationsListFragment : AbstractMainFragment() {
                 val displayChatRoom = sharedViewModel.displayedChatRoom
                 if (displayChatRoom != null) {
                     val found = listViewModel.conversations.value.orEmpty().find { model ->
-                        model.chatRoom == displayChatRoom
+                        model.conversationModel?.chatRoom == displayChatRoom
                     }
-                    found?.updateUnreadCount()
+                    found?.conversationModel?.updateUnreadCount()
                 }
                 listViewModel.updateUnreadMessagesCount()
             }
@@ -342,5 +404,22 @@ class ConversationsListFragment : AbstractMainFragment() {
         } catch (e: IllegalStateException) {
             Log.e("$TAG Failed to unregister data observer to adapter: $e")
         }
+    }
+
+    private fun showNumbersOrAddressesDialog(list: List<ContactNumberOrAddressModel>) {
+        val numberOrAddressModel = NumberOrAddressPickerDialogModel(list)
+        val dialog =
+            DialogUtils.getNumberOrAddressPickerDialog(
+                requireActivity(),
+                numberOrAddressModel
+            )
+
+        numberOrAddressModel.dismissEvent.observe(viewLifecycleOwner) { event ->
+            event.consume {
+                dialog.dismiss()
+            }
+        }
+
+        dialog.show()
     }
 }